Some eye candy

+++ test/matcher_full/main_matcher_full.cpp
@@ -86,7 +86,7 @@ bool checkFiles()
         std::cout << "Error: unable to find data dir" << std::endl;
         exit(0);
     }
-    for(size_t i = 0; i < N; ++i)
+    for(int i = 0; i < N; ++i)
     {
         std::stringstream ss;
         ss << base << "kermit" << std::setw(3) << std::setfill('0') << i << ".jpg";
